+++ CHANGELOG.md
@@ -42,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Disallow an empty string for a `StoreKey`
  - `ArrayBuilder` now validates additional fields
  - `FillValue::equals_all` incorrect behaviour with a `FillValue` with size not equal to 1, 2, 4, 8, or 16 bytes.
+ - Fix `NodePath` display output
 
 ## [0.10.0] - 2024-01-17